pyproject.toml
[tool.poetry]
name = "exampledjangocrud"
version = "0.1.0"
description = "Django CRUD RESTful API"
authors = ["Oleksandr Besan <oleksandr.besan@okiiyume.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
python-decouple = "^3.8"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
